totd image is not deleted if time is before 11pm

+++ bot/cogs/listeners/on_ready.py
@@ -65,9 +65,11 @@ class OnReady(
 
         # Deleting the TOTD Image if it exists
         if os.path.exists("./bot/resources/temp/totd.png"):
-            log.critical("TOTD Image Exists, Deleting")
-            os.remove("./bot/resources/temp/totd.png")
-
+            if int(datetime.datetime.now().hour) >= 23:
+                log.critical("TOTD Image Exists, Deleting")
+                os.remove("./bot/resources/temp/totd.png")
+            else:
+                log.critical('Time is not after 11pm, not deleting TOTD Image')
         # Starting TOTDImageDeleter
         log.info("Starting TOTDImageDeleter")
         totd_image_deleter.start(self.bot)
